feat(config): enhance script configuration handling and introduce new config interfaces
