+++ scripts/go/go-analyze-alarm/src/config.ts
@@ -45,15 +45,3 @@ export const scriptParameters: ReadonlyArray<Core.GOConfigParameterOptions> = [
   },
 ] as const;
 
-/**
- * Script configuration interface
- * Represents all validated configuration parameters
- */
-export interface GoAnalyzeAlarmConfig {
-  /** Name of the CloudWatch alarm */
-  readonly alarmName: string;
-  /** Timestamp when the alarm triggered (ISO 8601) */
-  readonly alarmDatetime: string;
-  /** AWS SSO profile names */
-  readonly awsProfiles: ReadonlyArray<string>;
-}
