+++ scripts/go/go-analyze-alarm/src/types/index.ts
@@ -1,9 +1,5 @@
 /**
  * Types barrel file
  * Re-export all types from this directory
- *
- * Usage: Create type files in PascalCase (e.g., User.ts, Config.ts)
- * then export them here:
- *   export type { User } from './User.js';
- *   export type { Config } from './Config.js';
  */
+export type { GoAnalyzeAlarmConfig } from './GoAnalyzeAlarmConfig.js';
