+++ scripts/send/send-dump-sqs/src/types/SendDumpSqsConfig.ts
@@ -1,0 +1,27 @@
+import { SendDumpSqsDedupMode } from './SendDumpSqsDedupMode.js';
+
+/**
+ * Script configuration interface
+ */
+export interface SendDumpSqsConfig {
+  /** AWS SSO profile name */
+  readonly awsProfile: string;
+
+  /** Target SQS queue name */
+  readonly queueName: string;
+
+  /** Visibility timeout for received messages */
+  readonly visibilityTimeout: number;
+
+  /** Maximum number of messages to dump */
+  readonly limit: number | undefined;
+
+  /** Deduplication mode */
+  readonly dedupMode: SendDumpSqsDedupMode;
+
+  /** Number of consecutive empty polls before stopping */
+  readonly maxEmptyReceives: number;
+
+  /** Custom output file path */
+  readonly outputFile: string | undefined;
+}
