+++ scripts/send/send-dump-sqs/src/types/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Types barrel file
+ * Re-export all types from this directory
+ */
+export { SendDumpSqsDedupMode } from './SendDumpSqsDedupMode.js';
+export type { SendDumpSqsConfig } from './SendDumpSqsConfig.js';
