+++ scripts/send/send-fetch-timeline-from-iun/src/types/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Types barrel file
+ * Re-export all types from this directory
+ */
+export type { SendFetchTimelineFromIunConfig } from './SendFetchTimelineFromIunConfig.js';
