+++ scripts/send/send-import-notifications/src/types/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Types barrel file
+ * Re-export all types from this directory
+ */
+export type { ImportNotificationsConfig } from './ImportNotificationsConfig.js';
